#243 refactored autoIssue to own service (origin: scheduler)
